move GoPublic to Corporation/Actions to support testability
